* commit '969a8d24fcdc648b3f88d28d47280869f4974396':
  :art: Improve web page clip code block content conversion https://github.com/siyuan-note/siyuan/issues/9896
  :bug: Create doc failed on non-Windows https://github.com/siyuan-note/siyuan/issues/9890
  :bug: fix https://github.com/siyuan-note/siyuan/issues/9850
  :art: Add Relation and Rollup column to database table view https://github.com/siyuan-note/siyuan/issues/9888
  :bug: Do not remove .tmp file when listing doc tree https://ld246.com/article/1702621572239
  :recycle: Refactor database table view
  :bug: fix https://github.com/siyuan-note/siyuan/issues/9891
  :art: delete row by keydown
  :art: fix https://github.com/siyuan-note/siyuan/issues/9886
  :art: https://github.com/siyuan-note/siyuan/issues/9886
  :art: Enable or disable all CSS and JS code snippets https://github.com/siyuan-note/siyuan/issues/9860
  :art: Database table view template columns support reference date columns https://github.com/siyuan-note/siyuan/issues/9887
  :art: fix https://github.com/siyuan-note/siyuan/issues/9878
  :arrow_up: Upgrade kernel deps
  :art: Delete the corresponding data in the database attributes table after deleting the doc https://github.com/siyuan-note/siyuan/issues/9875
  :arrow_up: Upgrade Lute
  :art: Add internal kernel API `/api/block/getTailChildBlocks` https://github.com/siyuan-note/siyuan/issues/9884
  :art: https://github.com/siyuan-note/siyuan/issues/9882
  :bug:

+++ kernel/model/box.go
@@ -258,10 +258,12 @@ func (box *Box) Ls(p string) (ret []*FileInfo, totals int, err error) {
 			continue
 		}
 		if strings.HasSuffix(name, ".tmp") {
-			// 移除写入失败时产生的临时文件
+			// 移除写入失败时产生的并且早于 30 分钟前的临时文件，近期创建的临时文件可能正在写入中
 			removePath := filepath.Join(util.DataDir, box.ID, p, name)
-			if removeErr := os.Remove(removePath); nil != removeErr {
-				logging.LogWarnf("remove tmp file [%s] failed: %s", removePath, removeErr)
+			if info.ModTime().Before(time.Now().Add(-30 * time.Minute)) {
+				if removeErr := os.Remove(removePath); nil != removeErr {
+					logging.LogWarnf("remove tmp file [%s] failed: %s", removePath, removeErr)
+				}
 			}
 			continue
 		}
